Moves angelfish page props -> home page props
