+++ app/services/homePagePropsService.ts
@@ -1,7 +1,10 @@
-import { HomePageProps } from "../pages";
+import { HomePageProps } from "../pages/index";
 import { DateTime } from "luxon";
 import getVaccineData from "./vaccineDataService";
-import { DhbPopulationDoseData } from "../types/VaccineDataTypes";
+import {
+  DhbRegionId,
+  DhbPopulationDoseDataWithRegion,
+} from "../types/VaccineDataTypes";
 import { CONSTANTS } from "./constants";
 
 // making it global so hopefully it gets cached by lambda
@@ -25,59 +28,53 @@ const isCacheExpired = (data: HomePageProps, ttl: number) =>
 async function generateHomePageProps(): Promise<HomePageProps> {
   const vaccineData = await getVaccineData();
 
-  const aucklandDhbNames = ["Auckland", "Counties Manukau", "Waitemata"];
-  const aucklandDhbsPopulationDoseData = vaccineData.vaccinationsPerDhb.filter(
-    (d) => aucklandDhbNames.includes(d.dhbName)
-  );
+  const aucklandDhbNames = ["Waitemata", "Auckland", "Counties Manukau"];
+  const northIslandDhbNames = aucklandDhbNames.concat([
+    "Northland",
+    "Waikato",
+    "Lakes",
+    "Bay of Plenty",
+    "Tairawhiti",
+    "Taranaki",
+    "Hawkes Bay",
+    "MidCentral",
+    "Whanganui",
+    "Capital and Coast",
+    "Hutt Valley",
+    "Wairarapa",
+  ]);
+  const southIslandDhbNames = [
+    "Nelson Marlborough",
+    "West Coast",
+    "Canterbury",
+    "South Canterbury",
+    "Southern",
+  ];
 
-  const combinedAucklandDhbsPopulationDoseData =
-    getCombinedAucklandDhbsPopulationDoseData(aucklandDhbsPopulationDoseData);
+  const mappedWithRegion: DhbPopulationDoseDataWithRegion[] =
+    vaccineData.vaccinationsPerDhb.map((dhb) => {
+      const regions: DhbRegionId[] = [];
+      if (aucklandDhbNames.includes(dhb.dhbName)) {
+        regions.push("auckland");
+      }
+      if (northIslandDhbNames.includes(dhb.dhbName)) {
+        regions.push("northIsland");
+      }
+      if (southIslandDhbNames.includes(dhb.dhbName)) {
+        regions.push("southIsland");
+      }
+      return {
+        ...dhb,
+        regionIds: regions,
+      };
+    });
+
   return {
-    allDhbsPopulationDoseData: vaccineData.vaccinationsPerDhb,
-    aucklandDhbsPopulationDoseData,
-    combinedAucklandDhbsPopulationDoseData,
+    allDhbsPopulationDoseData: mappedWithRegion,
     dataValidAsAtTimeUtc: vaccineData.dataValidAsAtTimeUtc.toISO(),
     dataFetchedAtTimeUtc: DateTime.now().toUTC().toISO(),
   };
 }
-
-const getCombinedAucklandDhbsPopulationDoseData = (
-  data: DhbPopulationDoseData[]
-): DhbPopulationDoseData => {
-  const combinedData: DhbPopulationDoseData = {
-    dhbName: "Combined Auckland DHBs",
-    numOfFirstDoses: 0,
-    firstDosesPercentage: 0,
-    hasMetFirstDoseTarget: false,
-    numOfFirstDosesTo90Percent: 0,
-    numOfSecondDoses: 0,
-    secondDosesPercentage: 0,
-    hasMetSecondDoseTarget: false,
-    numOfSecondDosesTo90Percent: 0,
-    totalPopulation: 0,
-  };
-
-  data.forEach((d) => {
-    combinedData.numOfFirstDoses += d.numOfFirstDoses;
-    combinedData.numOfFirstDosesTo90Percent += d.numOfFirstDosesTo90Percent;
-    combinedData.numOfSecondDoses += d.numOfSecondDoses;
-    combinedData.numOfSecondDosesTo90Percent += d.numOfSecondDosesTo90Percent;
-    combinedData.totalPopulation += d.totalPopulation;
-    combinedData.firstDosesPercentage = roundTo2Dp(
-      combinedData.numOfFirstDoses / combinedData.totalPopulation
-    );
-    combinedData.hasMetFirstDoseTarget =
-      combinedData.firstDosesPercentage >= CONSTANTS.firstDoseTargetPercentage;
-    combinedData.secondDosesPercentage = roundTo2Dp(
-      combinedData.numOfSecondDoses / combinedData.totalPopulation
-    );
-    combinedData.hasMetSecondDoseTarget =
-      combinedData.secondDosesPercentage >=
-      CONSTANTS.secondDoseTargetPercentage;
-  });
-
-  return combinedData;
-};
 
 const roundTo2Dp = (num: number) =>
   Math.round((num + Number.EPSILON) * 100) / 100;
